requirements.txt
# ==== Flask Web Arayüzü ====
Flask==3.1.0
Werkzeug==3.1.3
Jinja2==3.1.6
click==8.1.8
itsdangerous==2.2.0
blinker==1.9.0
importlib-metadata==8.7.0

# ==== Veri ve Görsel İşleme ====
numpy==1.26.4
pandas==2.2.3
Pillow==10.4.0
opencv-python==4.10.0.84
scikit-image==0.24.0
matplotlib==3.9.2
tqdm==4.67.1

# ==== Derin Öğrenme ====
torch==2.3.1
torchvision==0.18.1
torchaudio==2.3.1
timm==1.0.11  # MiDaS modeli için zorunlu

# ==== TensorFlow (Mac M2 Metal destekli) ====
tensorflow-macos==2.16.1
tensorflow-metal==1.1.0
keras==3.10.0

# ==== Yardımcı Kütüphaneler ====
protobuf==4.25.8
requests==2.32.5
absl-py==2.3.1
rich==14.2.0
h5py==3.14.0
